Fix Some Bangumi Track Bug (#4726)

+++ app/src/main/java/eu/kanade/tachiyomi/data/track/bangumi/Collection.kt
@@ -8,7 +8,7 @@ data class Collection(
     val comment: String? = "",
     val ep_status: Int? = 0,
     val lasttouch: Int? = 0,
-    val rating: Int? = 0,
+    val rating: Float? = 0f,
     val status: Status? = Status(),
     val tag: List<String?>? = listOf(),
     val user: User? = User(),
